+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 * Added classes for Germ Network lexicon.
 * Added classes for Standard.Site lexicon.
 * Added `Labels` property to `StatusView` to represent labels associated with a status. See [[APP-1775] Hydrate labels for actor statuses](https://github.com/bluesky-social/atproto/pull/4555)
-* Added `OpenGraphClient` to generate embedded records from Open Graph metadata, and documented its use.
+* Added `OpenGraphEmbeddedCardGenerator` to generate embedded records from Open Graph metadata, and documented its use.
 
 ## Changed